+++ src/main/java/com/netpoint/main/dto/AuthenticatedUser.java
@@ -1,4 +1,4 @@
 package com.netpoint.main.dto;
 
-public record AuthenticatedUser(String userId, String role, Long companyId) {
+public record AuthenticatedUser(String userId, String role, Integer companyId) {
 }
